requirements.txt
httpx[socks]
rich
requests
pysocks
pyfiglet
colorama
speedtest-cli